Move the sanitize fn too
