+++ apps/webapp/app/services/upsertBranch.server.ts
@@ -2,10 +2,10 @@ import { type PrismaClient, type PrismaClientOrTransaction } from "@trigger.dev/
 import slug from "slug";
 import { prisma } from "~/db.server";
 import { createApiKeyForEnv, createPkApiKeyForEnv } from "~/models/api-key.server";
+import { type CreateBranchOptions } from "~/routes/_app.orgs.$organizationSlug.projects.$projectParam.env.$envParam.branches/route";
+import { isValidGitBranchName, sanitizeBranchName } from "~/v3/gitBranch";
 import { logger } from "./logger.server";
 import { getLimit } from "./platform.v3.server";
-import { type CreateBranchOptions } from "~/routes/_app.orgs.$organizationSlug.projects.$projectParam.env.$envParam.branches/route";
-import { isValidGitBranchName } from "~/v3/validGitBranch";
 
 export class UpsertBranchService {
   #prismaClient: PrismaClient;
@@ -166,18 +166,4 @@ export async function checkBranchLimit(
     limit,
     isAtLimit: used >= limit,
   };
-}
-
-export function sanitizeBranchName(ref: string): string | null {
-  if (!ref) return null;
-  if (ref.startsWith("refs/heads/")) return ref.substring("refs/heads/".length);
-  if (ref.startsWith("refs/remotes/")) return ref.substring("refs/remotes/".length);
-  if (ref.startsWith("refs/tags/")) return ref.substring("refs/tags/".length);
-  if (ref.startsWith("refs/pull/")) return ref.substring("refs/pull/".length);
-  if (ref.startsWith("refs/merge/")) return ref.substring("refs/merge/".length);
-  if (ref.startsWith("refs/release/")) return ref.substring("refs/release/".length);
-  //unknown ref format, so reject
-  if (ref.startsWith("refs/")) return null;
-
-  return ref;
 }
