+++ apps/webapp/app/v3/gitBranch.ts
@@ -28,3 +28,17 @@ export function isValidGitBranchName(branch: string): boolean {
 
   return true;
 }
+
+export function sanitizeBranchName(ref: string): string | null {
+  if (!ref) return null;
+  if (ref.startsWith("refs/heads/")) return ref.substring("refs/heads/".length);
+  if (ref.startsWith("refs/remotes/")) return ref.substring("refs/remotes/".length);
+  if (ref.startsWith("refs/tags/")) return ref.substring("refs/tags/".length);
+  if (ref.startsWith("refs/pull/")) return ref.substring("refs/pull/".length);
+  if (ref.startsWith("refs/merge/")) return ref.substring("refs/merge/".length);
+  if (ref.startsWith("refs/release/")) return ref.substring("refs/release/".length);
+  //unknown ref format, so reject
+  if (ref.startsWith("refs/")) return null;
+
+  return ref;
+}
